backend/requirements.txt
Authlib==1.2.0
azure-functions==1.14.0